-dialogos de confirmacion para eliminar y finalizar proyectos -avisos generales enviados por el admin. -ventana de vista de proyectos adaptada por usuario

+++ app/src/main/java/com/example/davidblanco/projectod/ViewProjectFragment.kt
@@ -1,7 +1,9 @@
 package com.example.davidblanco.projectod
 
+import android.app.AlertDialog
 import android.app.FragmentManager
 import android.content.Context
+import android.content.SharedPreferences
 import android.net.Uri
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -12,37 +14,92 @@ import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import android.widget.Spinner
 import android.widget.Toast
+import com.google.android.gms.common.GooglePlayServicesNotAvailableException
 import com.google.firebase.database.FirebaseDatabase
+import kotlinx.android.synthetic.main.fragment_view_project.*
 import kotlinx.android.synthetic.main.fragment_view_project.view.*
+import android.content.DialogInterface
+import android.net.sip.SipSession
+import kotlinx.android.synthetic.main.comment_dialog.view.*
+import kotlinx.android.synthetic.main.project_dialog.view.*
 
 
 class ViewProjectFragment : Fragment(), View.OnClickListener {
     val database = FirebaseDatabase.getInstance()
     val myRef = database.getReference("Proyectos")
     val myRef2 = database.getReference("ProyectosReady")
+    val myRef3 = database.getReference("Avisos")
     var prj: Project = Project();
-    var key2:String="";
+    var key2: String = "";
+    var tipo: Int = 0;
     override fun onClick(v: View?) {
         val i = v!!.getId()
         if (i == R.id.btnCheck2) {
-            moveElements()
+            val builder = AlertDialog.Builder(activity)
+
+            builder.setTitle("Finalizar proyecto")
+                    .setMessage("¿El proyecto se ejecutó?, si desea puede agregar un comentario")
+            builder.setPositiveButton("SI") { dialog, which ->
+                moveElements()
+                dialog.dismiss()
+            }
+            builder.setNegativeButton("NO") { dialog, which ->
+                dialog.dismiss()
+            }
+            builder.setNeutralButton("Comentario") { dialog, which ->
+                val mDialogView = LayoutInflater.from(context).inflate(R.layout.comment_dialog, null)
+                val mBuilder = AlertDialog.Builder(context).setView(mDialogView)
+                val mAlertDialog = mBuilder.show()
+                mDialogView.btnSendComment.setOnClickListener {
+                    prj.comentario = mDialogView.etxtDescripcion2.text.toString()
+                    moveElements()
+                    mAlertDialog.dismiss()
+
+                }
+                mDialogView.btncancel2.setOnClickListener {
+                    mAlertDialog.dismiss()
+                    dialog.dismiss()
+                }
+            }
+            val alert = builder.create()
+            alert.show()
+
         } else if (i == R.id.btnDelete2) {
-            deleteElements(prj.key)
+            val builder = AlertDialog.Builder(activity)
+
+            builder.setTitle("Eliminar proyecto")
+                    .setMessage("¿Desea eliminar este proyecto de forma definitiva?")
+            builder.setPositiveButton("SI") { dialog, which ->
+                deleteElements(prj.key)
+                dialog.dismiss()
+            }
+            builder.setNegativeButton("NO") { dialog, which ->
+
+                dialog.dismiss()
+            }
+
+            val alert = builder.create()
+            alert.show()
+        } else if (i == R.id.btnVolver2) {
+            volverAtras()
         }
     }
 
     private fun moveElements() {
-        val key=myRef2.push().key
-        key2=prj.key
-        prj=Project(key!!,prj.titulo,prj.descripcion,prj.escuela,prj.fecha,prj.tipo)
+        val key = myRef2.push().key
+        key2 = prj.key
+        prj = Project(key!!, prj.titulo, prj.descripcion, prj.escuela, prj.fecha, prj.tipo, prj.email, prj.comentario)
+
         myRef2.child(key).setValue(prj)
         deleteElements(key2)
     }
 
-    private fun deleteElements(key:String) {
-        Toast.makeText(context, " " + prj.key,
-                Toast.LENGTH_SHORT).show()
-        myRef.child(key).removeValue()
+    private fun deleteElements(key: String) {
+        if(tipo==1) {
+            myRef.child(key).removeValue()
+        }else{
+            myRef3.child(key).removeValue()
+        }
         volverAtras();
     }
 
@@ -59,36 +116,65 @@ class ViewProjectFragment : Fragment(), View.OnClickListener {
         var view = inflater.inflate(R.layout.fragment_view_project, container, false)
         view.btnCheck2.setOnClickListener(this)
         view.btnDelete2.setOnClickListener(this)
+        view.btnVolver2.setOnClickListener(this)
         var bundle = getArguments()
-        prj = Project(bundle!!.getString("key"), bundle!!.getString("titulo"), bundle!!.getString("descripcion"), bundle!!.getString("escuela"),
-                bundle!!.getString("fecha"),bundle!!.getInt("tipo"))
+        prj = Project(bundle!!.getString("key"), bundle!!.getString("titulo"), bundle!!.getString("descripcion"),
+                bundle!!.getString("escuela"), bundle!!.getInt("fecha"), bundle!!.getInt("tipo"), bundle!!.getString("email"))
         view.txtViewT.setText(prj.titulo)
         view.txtViewD.setText(prj.descripcion)
-        view.txtSchoolV.setText(prj.escuela)
-        var spinner = view.findViewById(R.id.spLevel) as Spinner
-        var arrayAdapter = ArrayAdapter(context, android.R.layout.simple_spinner_item, arrayLevel)
-        arrayAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinner.adapter = arrayAdapter
-        spinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
-
-            override fun onNothingSelected(parent: AdapterView<*>?) {
-
+        view.txtSchoolV.setText("Escuela: " + prj.escuela)
+        view.txtFechaV.setText("Fecha: " + dateForrmat(prj.fecha))
+        mostrarDatos()
+        if (prj.escuela.equals("Departmento de educacion")) {
+            if(tipo == 0) {
+                view.btnDelete2.visibility = View.VISIBLE
+            }else {
+                view.btnDelete2.visibility = View.INVISIBLE
             }
-
-            override fun onItemSelected(parent: AdapterView<*>?, view: View?, position: Int, id: Long) {
-
-                if (position == 1) {
-                    prj.tipo = 1
-                } else if (position == 2) {
-                    prj.tipo = 2
-                } else if (position == 3) {
-                    prj.tipo = 3
+            view.btnCheck2.visibility = View.INVISIBLE
+            view.spLevel.visibility = View.INVISIBLE
+        } else if (tipo == 1) {
+            view.spLevel.visibility = View.INVISIBLE
+        } else {
+            view.btnDelete2.visibility = View.VISIBLE
+            view.btnCheck2.visibility = View.VISIBLE
+            view.spLevel.visibility = View.VISIBLE
+            var spinner = view.findViewById(R.id.spLevel) as Spinner
+            var arrayAdapter = ArrayAdapter(context, android.R.layout.simple_spinner_item, arrayLevel)
+            arrayAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+            spinner.adapter = arrayAdapter
+            spinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                override fun onNothingSelected(parent: AdapterView<*>?) {
                 }
-                myRef.child(prj.key).setValue(prj)
+                override fun onItemSelected(parent: AdapterView<*>?, view: View?, position: Int, id: Long) {
+                    if (position == 1) {
+                        prj.tipo = 1
+                    } else if (position == 2) {
+                        prj.tipo = 2
+                    } else if (position == 3) {
+                        prj.tipo = 3
+                    }
+                    myRef.child(prj.key).setValue(prj)
+                }
             }
         }
 
         return view
+    }
+
+    private fun dateForrmat(fecha: Int): String {
+        val cadena = "" + fecha
+        val d = cadena.toCharArray()
+        return cadena[6].toString() + cadena[7].toString() + "-" + cadena[4].toString() +
+                cadena[5].toString() + "-" + cadena[0].toString() + cadena[1].toString() +
+                cadena[2].toString() + cadena[3].toString();
+    }
+
+    private fun mostrarDatos() {
+        var sharedPreferences: SharedPreferences = this.activity!!.getSharedPreferences("Credenciales", Context.MODE_PRIVATE)
+        tipo = sharedPreferences.getInt("tipo", 1)
+
+
     }
